+++ backend/src/modules/logging/logging.module.ts
@@ -4,11 +4,12 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { LoggingController } from './controllers/logging.controller';
 import { SystemConfig } from './entities/system-config.entity';
 import { SystemLog } from './entities/system-log.entity';
+import { BranchConfig } from './entities/branch-config.entity';
 import { LoggingService } from './services/logging.service';
 import { LoggingCleanupService } from './services/logging-cleanup.service';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([SystemLog, SystemConfig])],
+  imports: [TypeOrmModule.forFeature([SystemLog, SystemConfig, BranchConfig])],
   providers: [LoggingService, LoggingCleanupService],
   controllers: [LoggingController],
   exports: [LoggingService],